Initial benchmarking example
