+++ libs/vm/benchmark/add_object_bench.cpp
@@ -1,0 +1,51 @@
+//------------------------------------------------------------------------------
+//
+//   Copyright 2018-2019 Fetch.AI Limited
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
+//
+//------------------------------------------------------------------------------
+
+#include "vm/module.hpp"
+#include "vm/vm.hpp"
+
+#include "benchmark/benchmark.h"
+
+#include <memory>
+
+#if 0
+namespace {
+
+void Add_Object_Benchmark(benchmark::State &state)
+{
+  auto module = std::make_shared<fetch::vm::Module>();
+
+  fetch::vm_modules::System::Bind(*module);
+  fetch::vm_modules::math::BindMath(*module);
+  fetch::vm_modules::ml::BindML(*module);
+  fetch::vm_modules::CreatePrint(*module);
+
+
+
+  fetch::vm::VM vm(module.get());
+
+  for (auto _ : state)
+  {
+    vm.
+  }
+}
+
+}  // namespace
+
+BENCHMARK(Add_Object_Benchmark);
+#endif
